requirements.txt

Django==5.0
djangorestframework==3.14.0


python-dateutil==2.8.2
pytz==2023.3.post1


sqlparse==0.4.4 
typing-extensions==4.8.0  
asgiref==3.7.2  

drf-yasg==1.21.7

django-debug-toolbar==4.2.0  
django-cors-headers==4.3.1 
